src/teleop_turbopi.cpp: Prepare for enum class, add static_cast<int>

+++ src/teleop_turbopi.cpp
@@ -29,7 +29,7 @@ namespace teleop_turbopi
     void TurboPi::joyCallback(const sensor_msgs::msg::Joy::SharedPtr joy_msg)
     {
         // system shutdown
-        if (joy_msg->buttons[TurboPi::buttons::SHARE])
+        if (joy_msg->buttons[static_cast<int>(TurboPi::buttons::SHARE)])
         {
             std::system("sudo init 0");
         }
@@ -38,8 +38,8 @@ namespace teleop_turbopi
         auto cmd_vel_msg = std::make_unique<geometry_msgs::msg::Twist>();
 
         // set wheel velocity
-        cmd_vel_msg->linear.x = joy_msg->axes[TurboPi::axes::LEFT_JOY_Y];
-        cmd_vel_msg->angular.z = joy_msg->axes[TurboPi::axes::LEFT_JOY_X];
+        cmd_vel_msg->linear.x = joy_msg->axes[static_cast<int>(TurboPi::axes::LEFT_JOY_Y)];
+        cmd_vel_msg->angular.z = joy_msg->axes[static_cast<int>(TurboPi::axes::LEFT_JOY_X)];
 
         // publish velocities
         publisher_cmd_vel_->publish(std::move(cmd_vel_msg));
@@ -49,9 +49,9 @@ namespace teleop_turbopi
         pos_msg->data.resize(2);
 
         // camera left/right
-        if (joy_msg->axes[TurboPi::axes::RIGHT_JOY_Y])
+        if (joy_msg->axes[static_cast<int>(TurboPi::axes::RIGHT_JOY_Y)])
         {
-            pos_msg->data[0] = joy_msg->axes[TurboPi::axes::RIGHT_JOY_Y];
+            pos_msg->data[0] = joy_msg->axes[static_cast<int>(TurboPi::axes::RIGHT_JOY_Y)];
         }
         else
         {
@@ -60,9 +60,9 @@ namespace teleop_turbopi
         }
 
         // camera up/down
-        if (joy_msg->axes[TurboPi::axes::RIGHT_JOY_X])
+        if (joy_msg->axes[static_cast<int>(TurboPi::axes::RIGHT_JOY_X)])
         {
-            pos_msg->data[1] = joy_msg->axes[TurboPi::axes::RIGHT_JOY_X];
+            pos_msg->data[1] = joy_msg->axes[static_cast<int>(TurboPi::axes::RIGHT_JOY_X)];
         }
         else
         {
